使用ndk r27进行编译

+++ ijkmedia/ijkplayer/android/pipeline/ffpipeline_android.c
@@ -111,7 +111,7 @@ static IJKFF_Pipenode *func_init_video_decoder(IJKFF_Pipeline *pipeline, FFPlaye
 static int func_config_video_decoder(IJKFF_Pipeline *pipeline, FFPlayer *ffp)
 {
     IJKFF_Pipeline_Opaque *opaque = pipeline->opaque;
-    int                       ret = NULL;
+    int                       ret = (int)NULL;
 
     if (ffp->node_vdec) {
         ret = ffpipenode_config_from_android_mediacodec(ffp, pipeline, opaque->weak_vout, ffp->node_vdec);
